Restores software MP3 decoding
